ft-seeders

+++ database/migrations/2024_10_09_051456_create_historiales_saldo_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('id_usuario') 
             ->constrained('clientes', 'id_usuario')
             ->onDelete('cascade');
-            $table->integer('id_historial_saldo');
+            $table->unsignedBigInteger('id_historial_saldo');
             $table->primary(['id_usuario', 'id_historial_saldo']);
             $table->timestamp('fecha_hora')->useCurrent();
             $table->double('monto');
